frontend/requirements.txt
streamlit
httpx
pillow
fastapi